+++ src/StringUtil.php
@@ -6,6 +6,8 @@ use Illuminate\Support\Str;
 
 class StringUtil
 {
+    public const WINDOWS_NEWLINE = "\r\n";
+
     /** https://en.wikipedia.org/wiki/Byte_order_mark#UTF-8 */
     public const UTF_8_BOM = "\xEF\xBB\xBF";
 
@@ -92,7 +94,7 @@ class StringUtil
         return \Safe\preg_split("/\r\n|\n|\r/", $string); // @phpstan-ignore return.type (preg_split from safe not known)
     }
 
-    public static function normalizeLineEndings(string $input, string $to = "\r\n"): string
+    public static function normalizeLineEndings(string $input, string $to = self::WINDOWS_NEWLINE): string
     {
         return \Safe\preg_replace("/\r\n|\r|\n/", $to, $input);
     }